S302T900-430 #time 50m Created delete buttons on news feed for all notification types. Fixed bug in listing like delete endpoint that occurred when removing a like from a listing that had no associated listing notification. Added missing documentation. ?implement ?document ?fix

+++ backend/src/main/java/org/seng302/controllers/ListingLikeController.java
@@ -122,7 +122,9 @@ public class ListingLikeController {
 
         //Remove old liked notification
         ListingNotification oldNotification = listingNotificationRepository.findListingNotificationsByUserIdAndListing(sessionUser.getId(), listing);
-        listingNotificationRepository.delete(oldNotification);
+        if (oldNotification != null) {
+            listingNotificationRepository.delete(oldNotification);
+        }
 
         //Add new liked notification
         ListingNotification notification = new ListingNotification(sessionUser, listing, NotificationStatus.UNLIKED);
